refactor: extract specific datacollection features from useDatasource to useDataCollectionSource
